+++ src/AbstractValue.php
@@ -1,18 +1,11 @@
 <?php
+
 /*
- * Copyright 2022 Cloud Creativity Limited
+ * Copyright 2025 Cloud Creativity Limited
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
  */
 
 declare(strict_types=1);
@@ -20,76 +13,86 @@ declare(strict_types=1);
 namespace CloudCreativity\Utils\Value;
 
 /**
- * Class AbstractValue
- *
- * @package CloudCreativity\Utils\Value
+ * @template TValue
+ * @implements ValueInterface<TValue>
  */
-abstract class AbstractValue implements ValueInterface
+abstract readonly class AbstractValue implements ValueInterface
 {
-
+    /** @use ValueTrait<TValue> */
     use ValueTrait;
 
     /**
      * Is the supplied scalar value acceptable for this value class?
-     *
-     * @param $value
-     * @return bool
      */
-    abstract protected function accept($value): bool;
+    abstract protected function accept(mixed $value): bool;
 
     /**
      * Fluent constructor.
-     *
-     * @param $value
-     * @return static
      */
-    public static function create($value): ValueInterface
+    public static function create(mixed $value): static
     {
-        return new static($value);
+        return static::from($value);
     }
 
     /**
-     * Cast the provided value.
-     *
-     * @param mixed $value
-     * @return static
+     * Create a value.
      */
-    public static function cast($value): ValueInterface
+    public static function from(mixed $value): static
     {
-        if ($value instanceof static) {
-            return $value;
-        }
-
-        if ($value instanceof ValueInterface) {
-            $value = $value->get();
-        }
-
-        return new static($value);
+        return match (true) {
+            $value instanceof static => $value,
+            $value instanceof ValueInterface => new static($value->get()),
+            default => new static($value),
+        };
     }
 
     /**
-     * Cast the provided value, if it is not null.
-     *
-     * @param mixed|null $value
-     * @return ValueInterface|null
+     * Create a value if it is acceptable, otherwise return null.
      */
-    public static function nullable($value): ?ValueInterface
+    public static function tryFrom(mixed $value): ?static
     {
-        if (is_null($value)) {
+        if ($value === null) {
             return null;
         }
 
-        return static::cast($value);
+        try {
+            return static::from($value);
+        } catch (ValueException) {
+            return null;
+        }
+    }
+
+    /**
+     * Backward compatibility method for casting values.
+     */
+    public static function cast(mixed $value): static
+    {
+        return static::from($value);
+    }
+
+    /**
+     * Cast the provided value if it is not null.
+     *
+     * @param mixed|null $value
+     */
+    public static function nullable(mixed $value): ?static
+    {
+        if ($value === null) {
+            return null;
+        }
+
+        return static::from($value);
     }
 
     /**
      * AbstractValue constructor.
      *
-     * @param mixed $value
      * @throws ValueException
      */
-    public function __construct($value)
+    final public function __construct(mixed $value)
     {
+        $value = $this->parse($value);
+
         if ($this->notAcceptable($value)) {
             throw new ValueException('Expecting a valid value.');
         }
@@ -98,12 +101,17 @@ abstract class AbstractValue implements ValueInterface
     }
 
     /**
-     * Is the value not acceptable?
-     *
-     * @param mixed $value
-     * @return bool
+     * Allow the value to be parsed before it is accepted.
      */
-    protected function notAcceptable($value): bool
+    protected function parse(mixed $value): mixed
+    {
+        return $value;
+    }
+
+    /**
+     * Is the value not acceptable?
+     */
+    protected function notAcceptable(mixed $value): bool
     {
         return !$this->accept($value);
     }