+++ src/ValueInterface.php
@@ -1,59 +1,41 @@
 <?php
+
 /*
- * Copyright 2022 Cloud Creativity Limited
+ * Copyright 2025 Cloud Creativity Limited
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
  */
 
 declare(strict_types=1);
 
 namespace CloudCreativity\Utils\Value;
 
+use BadMethodCallException;
 use JsonSerializable;
+use Stringable;
 
 /**
- * Interface ValueInterface
- *
- * @package CloudCreativity\Utils\Value
+ * @template TValue of mixed
  */
-interface ValueInterface extends JsonSerializable
+interface ValueInterface extends JsonSerializable, Stringable
 {
-
-    /**
-     * @return string
-     */
-    public function __toString();
-
     /**
      * Fluent string method.
-     *
-     * @return string
      */
     public function toString(): string;
 
     /**
-     * @return mixed
+     * @return TValue
      */
-    public function get();
+    public function get(): mixed;
 
     /**
      * Is the value any of the provided values?
      *
-     * @param mixed ...$values
-     * @return bool
-     * @throws \BadMethodCallException
-     *      if invoked without any values.
+     * @throws BadMethodCallException if invoked without any values.
      */
-    public function is(...$values): bool;
+    public function is(mixed ...$values): bool;
 
 }