+++ tests/Unit/ValueTest.php
@@ -1,24 +1,18 @@
 <?php
+
 /*
- * Copyright 2022 Cloud Creativity Limited
+ * Copyright 2025 Cloud Creativity Limited
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
  */
 
 declare(strict_types=1);
 
 namespace CloudCreativity\Utils\Value\Tests\Unit;
 
+use CloudCreativity\Utils\Value\Tests\IntegerMultiplierValue;
 use CloudCreativity\Utils\Value\Tests\IntegerValue;
 use CloudCreativity\Utils\Value\Tests\StringValue;
 use CloudCreativity\Utils\Value\ValueException;
@@ -27,7 +21,6 @@ use PHPUnit\Framework\TestCase;
 
 class ValueTest extends TestCase
 {
-
     public function testConstruct(): void
     {
         $value = new StringValue('abc');
@@ -121,5 +114,12 @@ class ValueTest extends TestCase
 
         $this->assertTrue($value->isNotEmpty());
         $this->assertFalse($value->isEmpty());
+    }
+
+    public function testItCanSetParser(): void
+    {
+        $value = new IntegerMultiplierValue(2);
+
+        $this->assertSame(20, $value->get());
     }
 }